// frontend/src/core/api/generated/model/shipmentRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */
import type { ShipmentLineRequest } from './shipmentLineRequest'

export interface ShipmentRequest {
  id: string
  /** @nullable */
  sourceId?: string | null
  /** @nullable */
  destinationId?: string | null
  /** @nullable */
  partyId?: string | null
  /**
   * @minLength 0
   * @maxLength 120
   * @nullable
   */
  partyName?: string | null
  /** @nullable */
  returnOfId?: string | null
  expectsReturn: boolean
  /** @nullable */
  day?: string | null
  /**
   * @minLength 0
   * @maxLength 500
   * @nullable
   */
  description?: string | null
  /** @minItems 1 */
  lines: ShipmentLineRequest[]
}